made adjustments to code.
