+++ config/config.json
@@ -1,7 +1,7 @@
 {
   "development": {
     "username": "root",
-    "password": "",
+    "password": "blue14",
     "database": "test_db",
     "host": "127.0.0.1",
     "port": 3306,
