+++ controllers/viewsController.js
@@ -1,9 +1,6 @@
 const express = require("express");
 const router = express.Router();
 
-router.get("/", (req, res) => {
-  res.render("index");
-});
 
 router.get("/index", (req, res) => {
   res.render("index");
@@ -25,7 +22,5 @@ router.get("/artist-post", (req, res) => {
   res.render("artist-post");
 });
 
-router.get("/trending", (req, res) => {
-  res.render("trending");
-});
+
 module.exports = router;
