+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "gt-project2",
   "version": "1.0.0",
-  "description": "",
+  "description": "blue14",
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
