+++ views/artist-dash.handlebars
@@ -51,7 +51,7 @@
   <footer>
     {{!-- <p>Footer</p> --}}
   </footer>
-</div>
+</div> --}}
