+++ views/layouts/main.handlebars
@@ -13,12 +13,12 @@
     <!-- Material Design Bootstrap -->
     <link href="https://cdnjs.cloudflare.com/ajax/libs/mdbootstrap/4.19.1/css/mdb.min.css" rel="stylesheet">
     <link href="../../public/stylesheets/style.css" rel="stylesheet">
-
 </head>
 
 <body>
     {{{body}}}
 </body>
+
 <!-- JQuery -->
 <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
 <!-- Bootstrap tooltips -->
