+++ views/login.handlebars
@@ -36,12 +36,14 @@
 <div class="container">
   <form class="text-center border border-light p-5" action="#!">
 
-    <p class="h4 mb-4">Login</p>
+    <p class="h4 mb-4" id="login-form">Login</p>
 
     <!-- Email -->
+    <label for="email">Email address</label>
     <input type="email" id="defaultLoginFormEmail" class="form-control mb-4" placeholder="Email">
 
     <!-- Password -->
+    <label for="password">Password</label>
     <input type="password" id="defaultLoginFormPassword" class="form-control mb-4" placeholder="Password">
 
     <!-- Sign in button -->
@@ -54,3 +56,19 @@
   </form>
 </div>
 <!-- Default form login -->
+
+<script>
+    $(document).ready(() => {
+        console.log("Hello world");
+        $("#login-form").on("submit", (event) => {
+            event.preventDefault();
+            const email = $("#email").val();
+            const password = $("#password").val();
+            $.post("/api/auth", { email, password })
+                .then(response => {
+                    console.log(response);
+                    window.location.replace(`/dashboard/${response.data.id}`)
+                })
+        })
+    })
+</script>
